+++ exporter/opentelemetry-exporter-jaeger-thrift/src/opentelemetry/exporter/jaeger/thrift/__init__.py
@@ -83,12 +83,12 @@ from typing import Optional
 from opentelemetry.exporter.jaeger.thrift.gen.jaeger import (
     Collector as jaeger_thrift,
 )
-from opentelemetry.exporter.jaeger.thrift.send.thrift import (
+from opentelemetry.exporter.jaeger.thrift.send import (
     AgentClientUDP,
     Collector,
 )
 from opentelemetry.exporter.jaeger.thrift.translate import Translate
-from opentelemetry.exporter.jaeger.thrift.translate.thrift import (
+from opentelemetry.exporter.jaeger.thrift.translate import (
     ThriftTranslator,
 )
 from opentelemetry.sdk.environment_variables import (
