+++ exporter/opentelemetry-exporter-jaeger-thrift/tests/test_jaeger_exporter_thrift.py
@@ -22,7 +22,7 @@ import opentelemetry.exporter.jaeger.thrift as jaeger_exporter
 from opentelemetry import trace as trace_api
 from opentelemetry.exporter.jaeger.thrift.gen.jaeger import ttypes as jaeger
 from opentelemetry.exporter.jaeger.thrift.translate import Translate
-from opentelemetry.exporter.jaeger.thrift.translate.thrift import (
+from opentelemetry.exporter.jaeger.thrift.translate import (
     ThriftTranslator,
 )
 from opentelemetry.sdk import trace
@@ -228,7 +228,9 @@ class TestJaegerExporter(unittest.TestCase):
 
         default_tags = [
             jaeger.Tag(
-                key="span.kind", vType=jaeger.TagType.STRING, vStr="internal",
+                key="span.kind",
+                vType=jaeger.TagType.STRING,
+                vStr="internal",
             ),
         ]
 
@@ -461,7 +463,8 @@ class TestJaegerExporter(unittest.TestCase):
         spans = translate._translate(ThriftTranslator())
 
         batch = jaeger.Batch(
-            spans=spans, process=jaeger.Process(serviceName="xxx"),
+            spans=spans,
+            process=jaeger.Process(serviceName="xxx"),
         )
 
         agent_client.emit(batch)
